chore: rename components folder to ui
