+++ apps/client/src/App.tsx
@@ -1,4 +1,4 @@
-import Layout from "./components/Layout";
+import Layout from "./ui/Layout";
 import GamePage from "./pages/game/GamePage";
 
 export function App() {
