+++ apps/client/src/ui/AppBar.tsx
